kind changes in trunk

+++ sample/FORTRAN/driver_sine_inplace.f90
@@ -36,11 +36,11 @@
       real(mytype), dimension(:,:,:),  allocatable :: B
       real(mytype) pi,twopi,sinyz,diff,cdiff,ccdiff,ans
 
-      integer(8) Ntot
+      integer(i8) Ntot
       real(mytype) factor
       real(mytype),dimension(:),allocatable:: sinx,siny,sinz
-      real(8) rtime1,rtime2,Nglob,prec
-      real(8) gt(4,3),gtcomm(3),tc
+      real(r8) rtime1,rtime2,Nglob,prec
+      real(r8) gt(4,3),gtcomm(3),tc
       integer ierr,nu,ndim,dims(2),nproc,proc_id
       integer istart(3),iend(3),isize(3)
       integer fstart(3),fend(3),fsize(3)
@@ -317,7 +317,7 @@
 
       use p3dfft
 
-      integer(8) nar,i
+      integer(i8) nar,i
       real(mytype) X(nar)
       real(mytype) f
 
@@ -337,8 +337,8 @@
       use p3dfft
 
       integer x,y,z,proc_id
-      integer(8) i,Nar
-      real(8) Nglob
+      integer(i8) i,Nar
+      real(r8) Nglob
       real(mytype), target :: Ar(1,1,*)
       integer Fstart(3),Fend(3),Fsize(3)
